pyproject.toml
[project]
name = "PlexMusicRatingsSync"
description = "PlexMusicRatingsSync keeps your Plex music ratings in sync with your audio files."
version = "0.0.0"
readme = "README.md"
authors = [{ name = "Ricardo Amaral", email = "hello@ricardoamaral.net" }]
license = { file = "LICENSE" }
requires-python = ">=3.8"
classifiers = [
    "Environment :: Console",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "click==8.1.8",
    "colorama==0.4.6",
    "mutagen==1.47.0",
    "platformdirs==4.3.6",
    "plexapi==4.16.1 ",
    "pyyaml==6.0.2",
]

[project.urls]
Repository = "https://github.com/rfgamaral/PlexMusicRatingsSync"
Issues = "https://github.com/rfgamaral/PlexMusicRatingsSync/issues"
Changelog = "https://github.com/rfgamaral/PlexMusicRatingsSync/blob/main/CHANGELOG.md"

[project.scripts]
plex-music-ratings-sync = "plex_music_ratings_sync.__main__:main"
pmrs = "plex_music_ratings_sync.__main__:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.semantic_release]
allow_zero_version = false
build_command = "pip install build && python -m build"
commit_parser = "conventional"
commit_parser_options = { parse_squash_commits = true, ignore_merge_commits = true }
logging_use_named_masks = true
mask_initial_release = true
version_variables = ["src/plex_music_ratings_sync/__init__.py:__version__"]
version_toml = ["pyproject.toml:project.version"]
